Generators for text blocks
