+++ generators/generators.js
@@ -27,48 +27,6 @@ import { registerFunctionsGenerators } from "./generators-functions.js";
 // Used outside of this file
 export * from "./mesh-state.js";
 
-function sanitizeForCode(input) {
-  let s = String(input);
-
-  // Cut from the first *real* newline (\r, \n, or Unicode line separator)
-  s = s.replace(/[\r\n\u2028\u2029].*$/s, "");
-  // Cut from the first *escaped* newline sequence (\n, \r, \u2028, \u2029, \x0A, \x0D)
-  s = s.replace(/\\(?:n|r|u(?:2028|2029|000a|000d)|x0(?:a|d)).*$/i, "");
-
-  // Remove any trailing backslashes that could remain (edge cases)
-  s = s.replace(/\\+$/, "");
-
-  // Neutralize comment and template literal markers
-  s = s.replace(/\*\//g, "*∕").replace(/\/\//g, "∕∕").replace(/`/g, "ˋ");
-
-  // Strip control characters (optional, keeps tabs/spaces)
-  s = s.replace(/[\u0000-\u001F\u007F]/g, "");
-
-  return s;
-}
-
-function emitSafeTextArg(code) {
-  if (!code) return '""';
-  const m = code.match(/^(['"`])(.*)\1$/s);
-  if (!m) return code;
-
-  const q = m[1];
-  const body = m[2];
-
-  // Decode literal safely (handles \', \\ , \n, \uXXXX, etc.)
-  let decoded;
-  try {
-    decoded = JSON.parse(q + body + q);
-  } catch {
-    decoded = body
-      .replace(/\\"/g, '"')
-      .replace(/\\'/g, "'")
-      .replace(/\\\\/g, "\\");
-  }
-
-  return JSON.stringify(sanitizeForCode(decoded));
-}
-
 export function defineGenerators() {
   // Allow Flock users to use "name" as a variable name
   const reservedWordsWithoutName = javascriptGenerator.RESERVED_WORDS_.split(
@@ -102,390 +60,6 @@ export function defineGenerators() {
 
     // Return the string value as a quoted literal
     return [`"${pivotOption}"`, javascriptGenerator.ORDER_ATOMIC];
-  };
-
-  // Assumes sanitizeForCode(text) is defined and in scope.
-  /**
-   * comment block -> single-line JS comment.
-   * Sanitizes the displayed text so it cannot break out of comment context.
-   */
-  javascriptGenerator.forBlock["comment"] = function (block) {
-    let raw =
-      javascriptGenerator.valueToCode(
-        block,
-        "COMMENT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "''";
-
-    const m = raw.match(/^(['"`])(.*)\1$/s);
-    const content = m ? m[2] : raw;
-
-    const safe = sanitizeForCode(content);
-    return `// ${safe}\n`;
-  };
-
-  javascriptGenerator.forBlock["describe"] = function (block) {
-    const textCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '""';
-
-    const meshVariable = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const safeTextArg = emitSafeTextArg(textCode);
-
-    return `await describeMesh(${meshVariable}, ${safeTextArg});\n`;
-  };
-
-  javascriptGenerator.forBlock["print_text"] = function (block) {
-    const textCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT",
-        javascriptGenerator.ORDER_NONE,
-      ) || "''";
-    const durationCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "DURATION",
-        javascriptGenerator.ORDER_NONE,
-      ) || "0";
-
-    const color = getFieldValue(block, "COLOR", '"#9932CC"');
-
-    const safeTextArg = emitSafeTextArg(textCode);
-
-    return `printText({ text: ${safeTextArg}, duration: ${durationCode}, color: ${color} });\n`;
-  };
-
-  javascriptGenerator.forBlock["ui_text"] = function (block) {
-    const textCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '""';
-    const xCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "X",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const yCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const fontSizeCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "FONT_SIZE",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "24";
-    const durationCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "DURATION",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const colorCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "COLOR",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '""';
-
-    const textBlockVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("TEXTBLOCK_VAR"),
-      Blockly.VARIABLE_CATEGORY_NAME,
-    );
-
-    const safeTextArg = emitSafeTextArg(textCode);
-
-    return `${textBlockVar} = UIText({
-          text: ${safeTextArg},
-          x: ${xCode},
-          y: ${yCode},
-          fontSize: ${fontSizeCode},
-          color: ${colorCode},
-          duration: ${durationCode},
-          id: ${textBlockVar}
-        });\n`;
-  };
-
-  javascriptGenerator.forBlock["ui_input"] = function (block) {
-    const varName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("INPUT_VAR"),
-      Blockly.VARIABLE_CATEGORY_NAME,
-    );
-
-    const textCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '""';
-    const xCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "X",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const yCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const fontSizeCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT_SIZE",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "24";
-    const textColorCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT_COLOR",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '"#000000"';
-    const backgroundColorCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "BACKGROUND_COLOR",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '"#ffffff"';
-
-    const size = block.getFieldValue("SIZE") || "medium";
-
-    const safeTextArg = emitSafeTextArg(textCode);
-
-    return `${varName} = await UIInput({
-            text: ${safeTextArg},
-            x: ${xCode},
-            y: ${yCode},
-            size: "${size}",
-            fontSize: ${fontSizeCode},
-            textColor: ${textColorCode},
-            backgroundColor: ${backgroundColorCode},
-            id: ${varName}
-          });\n`;
-  };
-
-  javascriptGenerator.forBlock["ui_slider"] = function (block) {
-    const varName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("SLIDER_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const min =
-      javascriptGenerator.valueToCode(
-        block,
-        "MIN",
-        javascriptGenerator.ORDER_NONE,
-      ) || 0;
-    const max =
-      javascriptGenerator.valueToCode(
-        block,
-        "MAX",
-        javascriptGenerator.ORDER_NONE,
-      ) || 100;
-    const value =
-      javascriptGenerator.valueToCode(
-        block,
-        "VALUE",
-        javascriptGenerator.ORDER_NONE,
-      ) || 50;
-    const x =
-      javascriptGenerator.valueToCode(
-        block,
-        "X",
-        javascriptGenerator.ORDER_NONE,
-      ) || 100;
-    const y =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y",
-        javascriptGenerator.ORDER_NONE,
-      ) || 50;
-    const color =
-      javascriptGenerator.valueToCode(
-        block,
-        "COLOR",
-        javascriptGenerator.ORDER_NONE,
-      ) || '"#000000"';
-    const background =
-      javascriptGenerator.valueToCode(
-        block,
-        "BACKGROUND",
-        javascriptGenerator.ORDER_NONE,
-      ) || '"#ffffff"';
-    const size = `"${block.getFieldValue("SIZE") || "MEDIUM"}"`;
-
-    const id = `"${varName}_slider"`;
-    const code = `
-                ${varName} = ${value};
-                const ${varName}_slider = UISlider({
-                  id: ${id},
-                  min: ${min},
-                  max: ${max},
-                  value: ${value},
-                  x: ${x},
-                  y: ${y},
-                  size: ${size},
-                  textColor: ${color},
-                  backgroundColor: ${background}
-                });
-                ${varName}_slider.onValueChangedObservable.add(value => {
-                  try { ${varName} = Math.round(value * 100) / 100; } catch (e) { console.warn('Variable not declared:', '${varName}'); }
-                });
-                `;
-
-    return code;
-  };
-
-  javascriptGenerator.forBlock["ui_button"] = function (block) {
-    // Retrieve values from the block
-    const text = javascriptGenerator.valueToCode(
-      block,
-      "TEXT",
-      javascriptGenerator.ORDER_ATOMIC,
-    );
-    const x = javascriptGenerator.valueToCode(
-      block,
-      "X",
-      javascriptGenerator.ORDER_ATOMIC,
-    );
-    const y = javascriptGenerator.valueToCode(
-      block,
-      "Y",
-      javascriptGenerator.ORDER_ATOMIC,
-    );
-    const width = `"${block.getFieldValue("SIZE")}"`; // Fix: Use "SIZE" instead of "WIDTH"
-    const textSize = `"${block.getFieldValue("TEXT_SIZE")}"`; // Fix: Add text size support
-    const textColor = javascriptGenerator.valueToCode(
-      block,
-      "TEXT_COLOR",
-      javascriptGenerator.ORDER_ATOMIC,
-    );
-    const backgroundColor = javascriptGenerator.valueToCode(
-      block,
-      "BACKGROUND_COLOR",
-      javascriptGenerator.ORDER_ATOMIC,
-    );
-
-    // Get the button variable
-    const buttonVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("BUTTON_VAR"),
-      Blockly.VARIABLE_CATEGORY_NAME,
-    );
-
-    const buttonId = `Button_${generateUniqueId()}`;
-
-    const code = `${buttonVar} = UIButton({
-                  text: ${text},
-                  x: ${x},
-                  y: ${y},
-                  width: ${width},
-                  textSize: ${textSize},
-                  textColor: ${textColor},
-                  backgroundColor: ${backgroundColor},
-                  buttonId: "${buttonId}"
-                });\n`;
-    return code;
-  };
-
-  javascriptGenerator.forBlock["say"] = function (block) {
-    const textCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "TEXT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || '""';
-    const durationCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "DURATION",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const alphaCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "ALPHA",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "1";
-    const sizeCode =
-      javascriptGenerator.valueToCode(
-        block,
-        "SIZE",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "24";
-
-    const meshVariable = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const textColor = getFieldValue(block, "TEXT_COLOR", "#000000");
-    const backgroundColor = getFieldValue(block, "BACKGROUND_COLOR", "#ffffff");
-
-    const mode = block.getFieldValue("MODE") || "";
-    const asyncMode = block.getFieldValue("ASYNC");
-    const asyncWrapper = asyncMode === "AWAIT" ? "await " : "";
-
-    const safeTextArg = emitSafeTextArg(textCode);
-
-    return `${asyncWrapper}say(${meshVariable}, { text: ${safeTextArg}, duration: ${durationCode}, textColor: ${textColor}, backgroundColor: ${backgroundColor}, alpha: ${alphaCode}, size: ${sizeCode}, mode: ${JSON.stringify(mode)} });\n`;
-  };
-
-  javascriptGenerator.forBlock["create_3d_text"] = function (block) {
-    const variableName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("ID_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    let rawText = getFieldValue(block, "TEXT", "Hello World");
-    if (typeof rawText !== "string") rawText = String(rawText ?? "");
-    const m = rawText.match(/^(['"`])(.*)\1$/s);
-    const textLiteral = JSON.stringify(sanitizeForCode(m ? m[2] : rawText));
-
-    const fontKey = block.getFieldValue("FONT") || "__fonts_FreeSans_Bold_json";
-    const size = getFieldValue(block, "SIZE", "50");
-    const depth = getFieldValue(block, "DEPTH", "1.0");
-    const x = getFieldValue(block, "X", "0");
-    const y = getFieldValue(block, "Y", "0");
-    const z = getFieldValue(block, "Z", "0");
-    const color = getFieldValue(block, "COLOR", '"#FFFFFF"');
-
-    let font = "./fonts/FreeSans_Bold.json";
-    if (fontKey === "__fonts_FreeSans_Bold_json")
-      font = "./fonts/FreeSans_Bold.json";
-
-    const meshId = "text_" + generateUniqueId();
-    meshMap[meshId] = block;
-    meshBlockIdMap[meshId] = block.id;
-
-    let doCode = "";
-    if (block.getInput("DO")) {
-      doCode = javascriptGenerator.statementToCode(block, "DO") || "";
-    }
-    doCode = doCode ? `async function() {\n${doCode}\n}` : "";
-
-    return `${variableName} = create3DText({
-          text: ${textLiteral},
-          font: '${font}',
-          color: ${color},
-          size: ${size},
-          depth: ${depth},
-          position: { x: ${x}, y: ${y}, z: ${z} },
-          modelId: '${meshId}'${doCode ? `,\n  callback: ${doCode}` : ""}
-        });\n`;
   };
 
   javascriptGenerator.forBlock["play_theme"] = function (block) {
@@ -757,14 +331,6 @@ export function defineGenerators() {
     const asyncWrapper = safeAsyncMode === "AWAIT" ? "await " : "";
 
     return `${asyncWrapper}speak(${meshVariable}, ${text}, { voice: "${voice}", rate: ${rate}, pitch: ${pitch}, volume: ${volume}, language: "${language}", mode: "${safeAsyncMode.toLowerCase()}" });\n`;
-  };
-
-  javascriptGenerator.forBlock["when_action_event"] = function (block) {
-    const action = block.getFieldValue("ACTION");
-    const event = block.getFieldValue("EVENT");
-    const statements_do = javascriptGenerator.statementToCode(block, "DO");
-
-    return `whenActionEvent("${action}", async () => {${statements_do}}, ${event === "ends"});\n`;
   };
 
   javascriptGenerator.forBlock["highlight"] = function (block) {
